backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AWS & Cloud
boto3==1.34.11
botocore==1.34.11

# LLM & ML
langchain==0.1.0
langchain-core==0.1.24

# Document Processing
PyPDF2==4.0.1
python-docx==0.8.11

# Utilities
requests==2.31.0
pydantic==2.5.2
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.1
